setup redux store
