+++ front-end/src/containers/DataTable.tsx
@@ -1,7 +1,10 @@
 import React from "react";
 import { Container, Button } from "@mui/material";
 import { DataGrid, GridColDef} from "@mui/x-data-grid";
+import { useSelector } from 'react-redux';
+import { generateAutoIds } from "../utils";
 
+//set columns
 const columns: GridColDef[] = [
   {
     field: "id",
@@ -72,95 +75,10 @@ const columns: GridColDef[] = [
   },
 ];
 
-const rows = [
-  {
-    name: "John",
-    gender: "male",
-    address: "Jon",
-    mobile: 715426257,
-    birthday: new Date("2023-06-29"),
-    age: 35,
-  },
-  {
-    name: "Alice",
-    gender: "female",
-    address: "Wonderland",
-    mobile: 715426257,
-    birthday: new Date("2023-07-19"),
-    age: 32,
-  },
-  {
-    name: "Michael",
-    gender: "male",
-    address: "New York",
-    mobile: 715426257,
-    birthday: new Date("2023-09-18"),
-    age: 38,
-  },
-  {
-    name: "Emma",
-    gender: "female",
-    address: "London",
-    mobile: 715426257,
-    birthday: new Date("2023-05-04"),
-    age: 27,
-  },
-  {
-    name: "Daniel",
-    gender: "male",
-    address: "Paris",
-    mobile: 715426257,
-    birthday: new Date("2023-03-04"),
-    age: 34,
-  },
-  {
-    name: "Olivia",
-    gender: "female",
-    address: "Los Angeles",
-    mobile: 715426257,
-    birthday: new Date("2023-12-14"),
-    age: 30,
-  },
-  {
-    name: "William",
-    gender: "male",
-    address: "Sydney",
-    mobile: 715426257,
-    birthday: new Date("2023-11-25"),
-    age: 41,
-  },
-  {
-    name: "Sophia",
-    gender: "female",
-    address: "Berlin",
-    mobile: 715426257,
-    birthday: new Date("2023-09-10"),
-    age: 24,
-  },
-  {
-    name: "Alexander",
-    gender: "male",
-    address: "Moscow",
-    mobile: 715426257,
-    birthday: new Date("2022-03-25"),
-    age: 36,
-  },
-];
-
-const generateAutoIds = (
-  rows: {
-    name: string;
-    gender: string;
-    address: string;
-    mobile: number;
-    birthday: Date;
-    age: number;
-  }[]
-) => {
-  return rows.map((row, index) => ({ ...row, id: index + 1 }));
-};
-
 function DataTable() {
+  //set rows
+  const rows = useSelector((state: {table: any; rows:any} ) => state.table.rows);
+
   return (
     <Container maxWidth="xl" sx={{display: "flex",justifyContent: "center",alignItems: "center",minHeight: "100vh",}}>
       <DataGrid
