+++ front-end/src/redux/rootStore.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import {tableReducer} from "./tableData/tableSlice";
+
+// Combine reducers
+const rootReducer = combineReducers({
+    table:tableReducer,
+});
+
+export default rootReducer;
