+++ front-end/src/redux/store.ts
@@ -1,0 +1,10 @@
+import { configureStore } from '@reduxjs/toolkit';
+import rootReducer from './rootStore';
+
+//configure store
+const store = configureStore({
+  reducer: rootReducer,
+});
+
+
+export default store;
