+++ front-end/src/utils/index.ts
@@ -1,0 +1,15 @@
+//generate IDs to rows
+const generateAutoIds = (
+    rows: {
+      name: string;
+      gender: string;
+      address: string;
+      mobile: number;
+      birthday: Date;
+      age: number;
+    }[]
+  ) => {
+    return rows.map((row, index) => ({ ...row, id: index + 1 }));
+  };
+
+export {generateAutoIds};
